requirements.txt
asgiref==3.11.0
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.2
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.9
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
packaging==25.0
pillow==12.0.0
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
service-identity==24.2.0
sqlparse==0.5.5
tomli==2.3.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.3
zope.interface==8.1.1
